[FEAT] OwnersRoot 생성(티켓 구매)

+++ src/main/java/com/example/demo/domain/ticket/service/TicketService.java
@@ -6,6 +6,6 @@ import java.util.List;
 public interface TicketService {
     void batchRegisterTicket(List<BigInteger> tokenIdList, List<String> cidList);
 
-    void completeTicket(String address, Long sessionId, Long tokenId);
+    void completeTicket(String address, Long sessionId, BigInteger tokenId);
 
 }
